\name{build}
\alias{build}
\title{Build Network}
\arguments{
  \item{object}{ReactomeFINetwork object.}

  \item{genes}{Character vector of gene names}

  \item{object}{ReactomeFINetwork object.}

  \item{genes}{Character vector of gene names}

  \item{object}{HotNet object.}

  \item{genescores}{data.frame of genes and scores}

  \item{delta}{}

  \item{fdr.threshold}{}

  \item{permutations}{}

  \item{auto.delta}{}
}
\value{
  ReactomeFINetwork ReactomeFINetwork object with fis
  attribute set

  ReactomeFINetwork ReactomeFINetwork object with fis
  attribute set

  HotNet Hotnet object with modules attribute set
}
\description{
  Build FI network from a list of genes.

  Build FI network from a list of genes.

  Build HotNet from a list of genes and scores.
}

