\name{extractFIs}
\alias{extractFIs}
\title{Extract FIs From XML Document
Extract FI gene pairs from an XML document returned from a server query.}
\usage{
extractFIs(doc)
}
\arguments{
  \item{doc}{XML document}
}
\value{
data.frame Data frame where each row represents an FI and
each of the two columns in the data frame contains a gene
involved in the FI.
}
\description{
Extract FIs From XML Document Extract FI gene pairs from an
XML document returned from a server query.
}

