\name{extractProteinInfo}
\alias{extractProteinInfo}
\title{extract Protein Info
Extract protein information including accession ID and DB name, protein
 name, and sequence.}
\usage{
extractProteinInfo(protein.node)
}
\arguments{
  \item{protein.node}{XML node containing protein
  information}
}
\value{
data.frame Data frame where each row corresponds to a
protein and the columns contain the information mentioned
above.
}
\description{
extract Protein Info Extract protein information including
accession ID and DB name, protein name, and sequence.
}

