\name{fis<-}
\alias{fis<-}
\title{Set FI Network Data}
\arguments{
  \item{object}{ReactomeFINetwork object.}

  \item{value}{Data frame representing the Reactome FI
  network data. Each row represents an interaction between
  two genes and consists of two columns - one for each
  gene.}
}
\value{
  ReactomeFINetwork
}
\description{
  Set the FI Network Data.
}

