\name{fis}
\alias{fis}
\title{Retrieve FI Network Data}
\arguments{
  \item{object}{ReactomeFINetwork object.}
}
\value{
  data.frame FI network data. Each row represents an
  interaction between two genes and consists of two columns
  - one for each gene.
}
\description{
  Retrieve the stored FI network interactions.
}

