% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactomeFI.R
\name{ld_reactomeFI}
\alias{ld_reactomeFI}
\title{Gene-Gene Function Interaction Annotation - Direction and Type of interaction}
\usage{
ld_reactomeFI(version)
}
\arguments{
\item{version}{2014, 2015}
}
\value{
a dataframe with 5 columns: gene1 gene1 Annotation Direction Score
}
\description{
Gene-Gene Function Interaction Annotation - Direction and Type of interaction
}
\examples{

## list the contents that are loaded into memory
 df <- ld_reactomeFI(version= 2015)

}
\keyword{Data}
\keyword{Reactome}
\keyword{annotation}

