\name{modules<-}
\alias{modules<-}
\title{Set Network Module Data}
\arguments{
  \item{object}{ReactomeFINetwork or HotNet object.}

  \item{value}{Network module information}
}
\description{
  Set network module data. ReactomeFINetwork objects accept
  data.frames and HotNet objects accept lists
}

