\name{plot}
\alias{plot}
\title{Plot Network}
\arguments{
  \item{object}{ReactomeFINetwork object}

  \item{color.modules}{Set to FALSE to turn off module
  colouring}
}
\value{
  ggplot ggplot object containing a visualization of the
  given network
}
\description{
  Plot the interaction network.
}

