\docType{class}
\name{ReactomeFIService-class}
\alias{ReactomeFIService}
\alias{ReactomeFIService-class}
\title{ReactomeFIService Class}
\usage{
ReactomeFIService(version = c("2009", "2012", "2013"))
}
\arguments{
  \item{version}{character version of Reactome FI network
  (2009, 2012, 2013)}
}
\value{
ReactomeFIService
}
\description{
Represents an interface to the ReactomeFI RESTful API.
Available versions include 2009, 2012, and 2013.

ReactomeFIService constructor.
}

