\docType{methods}
\name{service}
\alias{service}
\alias{service,HotNet-method}
\alias{service,ReactomeFINetwork-method}
\title{Retrieve ReactomeFIService Object}
\usage{
service(object)

\S4method{service}{HotNet}(object)

\S4method{service}{ReactomeFINetwork}(object)
}
\arguments{
  \item{object}{ReactomeFINetwork or HotNet object.}
}
\value{
ReactomeFIService
}
\description{
Retrieve the underlying ReactomeFIService object.
}

