\docType{methods}
\name{serviceURL}
\alias{serviceURL}
\alias{serviceURL,ReactomeFIService-method}
\title{RESTful API URL}
\usage{
serviceURL(object)

\S4method{serviceURL}{ReactomeFIService}(object)
}
\arguments{
  \item{object}{ReactomeFIService object.}
}
\value{
character
}
\description{
Retrieve the base URL for the ReactomeFI RESTful API.
}

