\docType{methods}
\name{subnet}
\alias{subnet}
\alias{subnet,HotNet-method}
\title{Create HotNet ReacomeFI Network}
\arguments{
  \item{object}{HotNet object.}

  \item{fdr}{False discovery rate threshold for HotNet
  modules}
}
\value{
  ReactomeFINetwork
}
\description{
  Create a ReactomeFI Network from a subset of the HotNet
  modules.
}

