\docType{methods}
\name{version}
\alias{version}
\alias{version,ReactomeFINetwork-method}
\alias{version,ReactomeFIService-method}
\title{Reactome FI Network Version (2009 or 2012)}
\arguments{
  \item{object}{ReactomeFIService object.}
}
\value{
  character
}
\description{
  Retrieve the Reactome FI network version.
}

