\docType{methods}
\name{build}
\alias{build}
\alias{build,ReactomeFINetwork,character-method}
\title{Build Network}
\arguments{
  \item{object}{ReactomeFINetwork object.}

  \item{genes}{Character vector of gene names}

  \item{object}{ReactomeFINetwork object.}

  \item{genes}{Character vector of gene names}
}
\value{
  ReactomeFINetwork ReactomeFINetwork object with fis
  attribute set

  ReactomeFINetwork ReactomeFINetwork object with fis
  attribute set
}
\description{
  Build FI network from a list of genes.

  Build FI network from a list of genes.
}

