\name{extractProteinInfo}
\alias{extractProteinInfo}
\title{extract Protein Info}
\usage{
  extractProteinInfo(protein.node)
}
\arguments{
  \item{protein.node}{XML node containing protein
  information}
}
\value{
  data.frame Data frame where each row corresponds to a
  protein and the columns contain the information mentioned
  above.
}
\description{
  Extract protein information including accession ID and DB
  name, protein name, and sequence.
}

