\docType{methods}
\name{modules}
\alias{modules}
\alias{modules,HotNet-method}
\alias{modules,ReactomeFINetwork-method}
\alias{modules<-}
\alias{modules<-,ReactomeFINetwork,data.frame-method}
\title{Retrieve Network Module Data}
\arguments{
  \item{object}{ReactomeFINetwork or HotNet object.}

  \item{object}{ReactomeFINetwork or HotNet object.}

  \item{value}{Network module information}
}
\value{
  data.frame
}
\description{
  Retreive network module data.

  Set network module data. ReactomeFINetwork objects accept
  data.frames and HotNet objects accept lists
}

