\name{ReactomeFINetwork}
\alias{ReactomeFINetwork}
\title{ReactomeFINetwork Class}
\usage{
  ReactomeFINetwork(version = c("2009", "2012"))
}
\arguments{
  \item{version}{Version of ReactomeFI network (2009 or
  2012).}
}
\value{
  ReactomeFINetwork ReactomeFINetwork S4 object with FI
  network generated from gene list.
}
\description{
  Represents a functional interaction network generated
  using the Reactome database and a given list of genes.

  ReactomeFINetwork constructor
}

