\name{ReactomeFIService}
\alias{ReactomeFIService}
\title{ReactomeFIService Class}
\usage{
  ReactomeFIService(version = "2009")
}
\arguments{
  \item{version}{character version of Reactome FI network
  (2009 or 2012)}
}
\value{
  ReactomeFIService
}
\description{
  Represents an interface to the ReactomeFI RESTful API.
  Available versions include 2009 and 2012

  ReactomeFIService constructor
}

