\docType{methods}
\name{service}
\alias{service}
\alias{service,HotNet-method}
\alias{service,ReactomeFINetwork-method}
\title{Retrieve ReactomeFIService Object}
\arguments{
  \item{object}{ReactomeFINetwork or HotNet object.}
}
\value{
  ReactomeFIService
}
\description{
  Retrieve the underlying ReactomeFIService object.
}

