\docType{methods}
\name{version}
\alias{version}
\alias{version,ReactomeFINetwork-method}
\alias{version,ReactomeFIService-method}
\title{Reactome FI Network Version}
\usage{
version(object)

\S4method{version}{ReactomeFINetwork}(object)

\S4method{version}{ReactomeFIService}(object)
}
\arguments{
  \item{object}{ReactomeFIService object.}
}
\value{
character
}
\description{
Retrieve the Reactome FI network version.
}

